% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucdiff.R
\name{aucdiff}
\alias{aucdiff}
\title{Fit a model for the difference of AUCs}
\usage{
aucdiff(
  w = NULL,
  d = NULL,
  beta,
  p.reduced,
  N = NULL,
  infl.fn,
  vcov.fn = NULL,
  hessian.fn = NULL,
  eps.finite.diff = 0.05,
  dgm.params
)
}
\arguments{
\item{w}{covariates}

\item{d}{class labels, 0 or 1}

\item{p.reduced}{the number of initial components of w used in the reduced model}

\item{N}{the total sample size, taken to be nrow(w) if w isn't NULL}

\item{infl.fn}{the influence function to use; see details for specification}

\item{vcov.fn}{a function returning the variance-covariance matrix of the coefficient estimates and hoeffding gradient; see details for specification}

\item{hessian.fn}{a function returning the hessian of the AUC; see details for specification}

\item{eps.finite.diff}{the step size for numeric differentiation of the Hessian (if hessian.fn is NULL) and Hoeffding gradient}

\item{dgm.params}{parameters of the data-generating mechanism to be passed to vov.fn and hessian.fn}
}
\value{
an object of class "aucdiff"
}
\description{
Fit a model for the difference of AUCs
}
