% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucdiff.R
\name{coefs.lda}
\alias{coefs.lda}
\title{Compute LDA coefficients}
\usage{
coefs.lda(x.0, x.1, params = NULL)
}
\arguments{
\item{x.0}{control observations}

\item{x.1}{case observations}

\item{params}{parameters of the normal model, which are estimated
if NULL value is passed}
}
\value{
coefficient estimates
}
\description{
Compute LDA coefficients
}
