% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucdiff.R
\name{confint.aucdiff}
\alias{confint.aucdiff}
\title{An approximate CI for the difference of AUCs}
\usage{
\method{confint}{aucdiff}(fit, level = 0.95, mc.reps = 5000)
}
\arguments{
\item{fit}{an object of class "aucdiff"}

\item{level}{level of the CI}

\item{mc.reps}{the number of samples used to estimate the quantile}
}
\value{
a pair of numbers giving the lower and upper endpoints of
the CI
}
\description{
An approximate CI for the difference of AUCs
}
