% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucdiff.R
\name{infl.glm}
\alias{infl.glm}
\title{Influence function for GLM estimates}
\usage{
infl.glm(x, d, params, terms.only = TRUE)
}
\arguments{
\item{d}{class labels, 0 or 1}

\item{params}{parameters of the GLM, including the link function
and its first and second derivatives}

\item{terms.only}{boolean, whether to return the terms of the
influence function directly or a class-weighted average}

\item{w}{covariates}
}
\value{
depending on "terms.only", either the terms of the
influence function directly or a class-weighted average
}
\description{
Influence function for GLM estimates
}
