% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucdiff.R
\name{infl.lda}
\alias{infl.lda}
\title{Influence function for LDA estimates}
\usage{
infl.lda(w, d, params = NULL, terms.only = TRUE)
}
\arguments{
\item{w}{covariates}

\item{d}{class labels, 0 or 1}

\item{params}{parameters of the normal model, which are estimated
if NULL value is passed}

\item{terms.only}{boolean, whether to return the terms of the
influence function directly or a class-weighted average}
}
\value{
depending on "terms.only", either the terms of the
influence function directly or a class-weighted average
}
\description{
Influence function for LDA estimates
}
