% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucdiff.R
\name{paucdiff}
\alias{paucdiff}
\title{Approximate CDF of AUC differences}
\usage{
paucdiff(q, fit, mc.reps = 5000)
}
\arguments{
\item{q}{quantile}

\item{fit}{an object of class "aucdiff"}

\item{mc.reps}{the number of samples used to estimate the CDF}
}
\value{
the approximate probability of an AUC difference being less than or equal to q
}
\description{
Approximate CDF of AUC differences
}
