% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucdiff.R
\name{plot.aucdiff}
\alias{plot.aucdiff}
\title{Plot the approximate CDF of the difference of AUCs}
\usage{
\method{plot}{aucdiff}(fit, mc.reps = 5000, ...)
}
\arguments{
\item{fit}{an object of class "aucdiff"}

\item{mc.reps}{the number of samples used to estimate the CDF}

\item{...}{additional parameters passed down to the graphics
routine}
}
\value{
side effect
}
\description{
Plot the approximate CDF of the difference of AUCs
}
