% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucdiff.R
\name{qaucdiff}
\alias{qaucdiff}
\title{Approximate quantile function of AUC differences}
\usage{
qaucdiff(p, fit, mc.reps = 5000)
}
\arguments{
\item{p}{probability}

\item{fit}{an object of class "aucdiff"}

\item{mc.reps}{the number of samples used to estimate the quantile}
}
\value{
an approximation to the number q such that the probability of an AUC difference being less than or equal to q is p
}
\description{
Approximate quantile function of AUC differences
}
