% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucdiff.R
\name{raucdiff}
\alias{raucdiff}
\title{Generate a random sample of AUC differences}
\usage{
raucdiff(n, fit)
}
\arguments{
\item{n}{the number of samples}

\item{fit}{an object of class "aucdiff"}
}
\value{
numeric vector consisting of samples
}
\description{
Generate a random sample of AUC differences
}
