% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucdiff.R
\name{var.normal.lda}
\alias{var.normal.lda}
\title{Variance-covariance matrix for LDA influence functions and Hoeffding gradient}
\usage{
var.normal.lda(beta.star, params)
}
\arguments{
\item{beta.star}{the probability limit of the LDA estimates}

\item{params}{parameters of the normal model, which are estimated
if NULL value is passed}
}
\value{
the variance-covariance matrix of the reduced coefficients, full coefficients, and Hoeffding gradient
}
\description{
Variance-covariance matrix for LDA influence functions and Hoeffding gradient
}
